Backend/requirements.txt
aiohttp
fastembed
Flask
flask_cors
huggingface-hub
llama-cloud
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-fastembed
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-openai
llama-index-llms-huggingface
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-vector-stores-qdrant
tqdm
pyjwt
numpy
openai
packaging
supabase
pandas
python-dotenv
gunicorn
qdrant-client